feat: 드롭다운 컴포넌트 분리 및 모달 외부 클릭 시 닫힘 기능 수정
- 드롭다운 UI 컴포넌트화 및 재사용 구조 정리
- 모달 배경 클릭 시 닫히지 않던 문제 해결
- axios baseURL을 환경변수로 설정하여 배포 대응
