+++ frontend/src/index.css
@@ -1,1 +1,5 @@
 @import 'tailwindcss';
+
+button {
+  cursor: pointer;
+}
