+++ frontend/src/lib/apiClient.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const apiClient = axios.create({
-  baseURL: 'http://localhost:3000/api',
+  baseURL: import.meta.env.VITE_BASE_URL,
   withCredentials: true,
 });
